Mock seeder CVV bug fix

+++ webservice/src/main/java/com/happymall/webservice/mockserver/MockSeeder.java
@@ -56,7 +56,7 @@ public class MockSeeder {
 	    			mock.setNameOnCard(nextRecord[1]);
 	    			mock.setValue(Double.parseDouble(nextRecord[2]));
 	    			mock.setRemainingValue(Double.parseDouble(nextRecord[3]));
-	    			String cvv = (Math.random() *1000) + "";
+	    			String cvv = (int)(Math.random() *1000) + "";
 	    			mock.setCvv(cvv);
 	    			mock.setIssuedDate(new Date());
 	    			mock.setExpiredDate(new Date());
